ugc_mongo_service/requirements.txt
motor==3.1.1
fastapi==0.85.1
python-dotenv==0.21.0
gunicorn==20.1.0
uvicorn==0.17.6
orjson==3.8.0
PyJWT==2.6.0
python-logstash-async==2.5.0